Moved the function we created to read json files/streams to general.py. Add some comments
